Add target_id in fuzzer (#14)
* Revamping results db interface a bit

Generalizing more.

* bump athena image in debug deployment. sanity checks passing

* Get rid of keep_snapshot flag

This flad was added so that we only keep snapshots for routes when an
exception was raised.  However, now that we are deduping exceptions, an
exception is only returned once when it is unique. And because we send
multiple requests for a route, the keep_snapshot flag gets overwritten.
So, lets just always keep snapshots in debug mode.

* results db is nammed athena

* add target id field to exceptions table

* passed sanity checks